Moved to GWT 2.4.0.
Bug fix: declutter was not putting the labels over the cells that
had been marked as used so this could cause labels to overlap (rarely).
Created a nice way to render the declutter output in the unit test to
a png file and then compare the png file to a known good one as
the validation step for some of the declutter unit tests. Refactored
out an "IconEngine" that is responsible for placing the icons
after declutter.

Bug fix: Undid some IE7 performance tweaks because these made tiles not
get loaded sometimes (need setUrl on image all the time).

Bug fix: When a layer goes from active=true to active=false,
we need to hide its images.

Reorged stats collection to be in the method for updateView.

Removed unused image pixsize from Icon.